requirements.txt
Django==4.2.7
Pillow==10.1.0
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
numpy==1.25.2
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
